Initial code
This is the whole WIP for the current approach where:
PLC1:
controls Water Level and Chlorine Pump
Communicates with PLC2

PLC2:
controls water motor
receive feedback if the watermotor is on.
Communicates watermotor status to PLC1

Expected workflow of the system:
Water Level is LOW
Motor is turned ON
Motor signal is LOW
Chlorine is turned ON.

All input ports are Grounded, meaning ON -> Reads LOW (Uses pullup).